--- source/ScrumblesFrames.py
@@ -1,3 +1,0 @@
-from SDynamics import *
-from SGraphs import *
-from SLists import *
