+++ source/frames/ScrumblesObjects.py
