+++ source/styling/tkColors.py
