+++ source/testing/sprintUnitTest.py
@@ -1,7 +1,8 @@
 ### test creation and deletion of a Sprint object
-import ScrumblesObjects
+from frames import ScrumblesObjects
 import datetime
-import ScrumblesData
+from data import ScrumblesData
+
 
 def testSprints(dataConnection):
     getAllSprintsQuery = ScrumblesData.Query.getAllSprints
