+++ source/testing/testSprintDelete.py
@@ -1,5 +1,5 @@
-import DataBlock
-import ScrumblesObjects
+from data import DataBlock
+from frames import ScrumblesObjects
 from datetime import datetime
 import time
 import logging
