+++ source/views/analyticsView.py
@@ -1,7 +1,7 @@
 import tkinter as tk
 from tkinter import ttk
 import listboxEventHandler
-import ScrumblesFrames
+from frames import ScrumblesFrames
 import sys
 from styling import styling as style
